more control
need retest though

+++ doprep/bootstrap.ps1
@@ -1,6 +1,9 @@
-﻿Get-Volume | ? { (($_.DriveType -eq "CD-ROM") -and ($_.FileSystem -eq "UDF")) } | % {
+﻿$myself = (Get-Variable MyInvocation).value.InvocationName
+Get-Volume | ? { (($_.DriveType -eq "CD-ROM") -and ($_.FileSystem -eq "UDF")) } | % {
     $doprep = ( "{0}:\doprep.ps1" -f $_.driveLetter)
-    & $doprep
+    if(Test-Path $doprep) {
+        mv $myself ("{0}.bak" -f $myself) 
+        & $doprep
+    }
 }
-$myself = (Get-Variable MyInvocation).value.InvocationName
-mv $myself ("{0}.bak" -f $myself) 
+
